# starter-resource-server/src/generated/swagger.yml
swagger: "2.0"
info:
    version: "0.0.1"
    title: "Swagger API"
host: localhost:4266
basePath: "/"
schemes:
    - http
    - https
consumes:
    - application/json
produces:
    - application/json
definitions:
...
    VideoCreateBody:
        type: object
        properties:
            name:
                type: string
            json:
                type: object
    VideoChangesBody:
        type: object
        properties:
            json:
                type: object
    VideoPushBody:
        type: object
    VideoPullBody:
        type: object
    VideoUpdateBody:
        type: object
        properties:
            id:
                type: string
            changes:
                $ref: '#/definitions/VideoChangesBody'
            push:
                $ref: '#/definitions/VideoPushBody'
            pull:
                $ref: '#/definitions/VideoPullBody'
    AddIdBody:
        type: object
        properties:
            addId:
                type: string
...
    RemoveIdBody:
        type: object
        properties:
            addId:
                type: string
...
paths:
...
    /stores/{id}/videos:
        get:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
    /videos:
        get:
            description: "sample description"
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
        post:
            description: "sample description"
            parameters:
                - name: body
                  description: "Not provided"
                  in: body
                  schema:
                    $ref: '#/definitions/VideoCreateBody'
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
    /videos/{id}:
        get:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
        put:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string

                - name: body
                  description: "Not provided"
                  in: body
                  schema:
                    $ref: '#/definitions/VideoUpdateBody'
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
        delete:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
    /videos/{id}/store:
        get:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
    /videos/{id}/store/add:
        put:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string

                - name: body
                  description: "Not provided"
                  in: body
                  schema:
                    $ref: '#/definitions/AddIdBody'
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
    /videos/{id}/store/remove:
        put:
            description: "sample description"
            parameters:
                - name: id
                  in: path
                  required: true
                  description: "Not provided"
                  type: string

                - name: body
                  description: "Not provided"
                  in: body
                  schema:
                    $ref: '#/definitions/RemoveIdBody'
            responses:
                '200':
                    description: "not provided."
                    schema:
                        type: object
...